Try switching to actors
